backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
slowapi>=0.1.9
